+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioned sections should match the Git tags and GitHub releases published for t
 
 ## [Unreleased]
 
+### Dependencies
+
+- Upgrade the tested Pi package contract from 0.84.3 to 0.85.0 ([#459](https://github.com/ladislas/mypac/issues/459)).
+
 ## [1.0.0] - 2026-09-03
 
 ### Added